moved react components into folder, installed firebase node module
