+++ client/package.json
@@ -4,8 +4,10 @@
   "private": true,
   "proxy": "http://localhost:3001/",
   "dependencies": {
+    "firebase": "^5.5.8",
     "react": "^16.0.0",
     "react-dom": "^16.0.0",
+    "react-router-dom": "^4.3.1",
     "react-scripts": "1.0.14"
   },
   "scripts": {
